+++ src/de/omegazirkel/risingworld/tools/ui/OZUIElement.java
@@ -1,11 +1,16 @@
 package de.omegazirkel.risingworld.tools.ui;
 
+import java.util.function.Consumer;
+
+import net.risingworld.api.events.player.ui.PlayerUIElementClickEvent;
 import net.risingworld.api.ui.UIElement;
 
 /**
  * extends UIElement with some helper methods missing from UIElement
  */
 public class OZUIElement extends UIElement {
+    protected Consumer<PlayerUIElementClickEvent> clickAction;
+
     public void setHoverBorderColor(int rgba) {
         hoverStyle.borderBottomColor.set(rgba);
         hoverStyle.borderLeftColor.set(rgba);
@@ -22,5 +27,16 @@ public class OZUIElement extends UIElement {
 
     public void setHoverBackgroundColor(int rgba) {
         hoverStyle.backgroundColor.set(rgba);
+    }
+
+    public void setClickAction(Consumer<PlayerUIElementClickEvent> clickAction) {
+        this.clickAction = clickAction;
+    }
+
+    @Override
+    public void onClick(PlayerUIElementClickEvent event) {
+        if (clickAction != null) {
+            clickAction.accept(event);
+        }
     }
 }